Replace oauth2 with rauth library for twitter.
The oauth2 library used by Syte [1] is an abandoned fork that has not
been updated in 3 years. It must be installed from github through git,
instead of a PyPi package. This ment it must be re-installed every
time one push changes to heroku.

This changeset replaces the oauth2 library with the rauth library,
which is maintained and has full test coverage [2].

* [1] https://github.com/brosner/python-oauth2
* [2] https://github.com/litl/rauth